refactor: split StateManager into StateProjectionService + InternalEventBus subscribers (M5)
- Rename StateManager → StateProjectionService in new file; old file is a
  compatibility re-export (StateManager alias) for downstream imports.
- Route all state-cache updates through InternalEventBus subscriptions instead
  of direct DaemonHub listeners:
  • session.created, session.updated, session.deleted
  • commands.updated, session.error, session.errorClear
  • api.connection, settings.updated
- Add DaemonInternalEventMap entries for SessionEvents and ApiConnectionEvents.
- Bridge migrated DaemonHub events to InternalEventBus in app.ts so existing
  publishers continue to work without touching every call site.
- Client delivery remains fully routed through ClientEventBridge (M4).
- Side effects (broadcastSystemChange, broadcastSettingsChange,
  broadcastSessionStateChange) are triggered by InternalEventBus subscribers
  inside StateProjectionService, keeping them separate from cache updates.
- Update client-event-bridge docs to reference StateProjectionService.
- Add knip ignore for the compatibility re-export file.
- All 9824 daemon unit tests pass; lint + typecheck + knip pass.
