+++ knip.json
@@ -30,6 +30,7 @@
 		"packages/daemon/src/lib/job-queue-constants.ts",
 		"packages/daemon/dump-context.ts",
 		"packages/daemon/src/lib/external-events/index.ts",
+		"packages/daemon/src/lib/state-manager.ts",
 		".claude/**"
 	],
 	"ignoreDependencies": [
